update: code view content and data handling split
